refactor sahp unit test
